+++ package.json
@@ -9,6 +9,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-icons": "^3.11.0",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "3.4.3",
     "ui-neumorphism": "^1.0.9"
   },
